+++ frontend/src/App.tsx
@@ -35,6 +35,7 @@ import { StatusTicketForm } from './pages/StatusTicketForm'
 import { StatusTicketDetalhe } from './pages/StatusTicketDetalhe'
 import { RespostasProntasPage } from './pages/RespostasProntas'
 import { RoteamentoRegrasPage } from './pages/RoteamentoRegras'
+import { SlaPoliticasPage } from './pages/SlaPoliticas'
 import { RespostaProntaForm } from './pages/RespostaProntaForm'
 import { RespostaProntaDetalhe } from './pages/RespostaProntaDetalhe'
 import { Auditoria } from './pages/Auditoria'
@@ -378,6 +379,7 @@ function AppRoutes() {
           <Route path="natureza-motivo" element={<TicketNaturezaMotivoPage embedded />} />
           <Route path="respostas-prontas" element={<RespostasProntasPage embedded />} />
           <Route path="roteamento" element={<RoteamentoRegrasPage embedded />} />
+          <Route path="sla" element={<SlaPoliticasPage embedded />} />
         </Route>
         <Route
           path="configuracoes/cadastros"